refactor: modularize client main file
