+++ client/src/main.ts
@@ -1,476 +1,45 @@
-import { assignElevations, assignIslandElevations, ElevationConfig } from './terrain-gen/elevations';
-import { assignBiomes, BiomeConfig, getBiomeName } from './terrain-gen/biomes';
+import { MapSize, deserializeTypedArrays } from './mesh';
 import { drawCells } from './drawCells';
-import { loadMesh, preloadAllMeshes, MapSize, MeshData, deserializeTypedArrays } from './mesh';
+import { createUI } from './ui';
+import {
+  loadOrGetMesh,
+  preloadMeshes,
+  currentMapSize,
+  currentCellBiomes,
+  currentCellCount,
+  showError,
+  meshData,
+  biomeConfig,
+  setCurrentCellBiomes,
+  setCurrentCellCount,
+  setCurrentMapSize,
+} from './terrain';
+import { WIDTH, HEIGHT, SERVER_BASE_URL } from './config';
 
-/**
- * Loads or waits for mesh data for a specific size
- */
-export async function loadOrGetMesh(size: MapSize): Promise<void> {
-  console.time('loadOrGetMesh');
+const canvas = document.createElement('canvas');
+const container = document.getElementById('canvas-container')!;
+container.appendChild(canvas);
+canvas.width = WIDTH;
+canvas.height = HEIGHT;
 
-  const fetchedMesh = await loadMesh(size, SERVER_BASE_URL);
-
-  if (fetchedMesh) {
-    meshData = fetchedMesh;
-
-    currentCellCount = meshData.cellCount;
-    if (currentCellBiomes.length !== currentCellCount) {
-      currentCellBiomes = new Uint8Array(currentCellCount);
-    }
-
-    generateTerrain();
-  } else {
-    showError(`Failed to load ${size} mesh from server`);
-  }
-
-  console.timeEnd('loadOrGetMesh');
-}
-
-/**
- * Shows an error message
- */
-function showError(message: string): void {
-  let errorElement = document.getElementById('errorMessage');
-  if (!errorElement) {
-    errorElement = document.createElement('div');
-    errorElement.id = 'errorMessage';
-    errorElement.style.cssText = `
-      position: fixed;
-      top: 20px;
-      left: 50%;
-      transform: translateX(-50%);
-      background: #f44336;
-      color: white;
-      padding: 10px 20px;
-      border-radius: 4px;
-      font-family: Arial, sans-serif;
-      z-index: 2000;
-    `;
-    document.body.appendChild(errorElement);
-  }
-  errorElement.textContent = message;
-  errorElement.style.display = 'block';
-}
-
-export function generateTerrain() {
-  if (!meshData) {
-    console.warn('No mesh data available for terrain generation');
-    return;
-  }
-
-  if (!meshData.cellTriangleCenters) {
-    console.warn('No triangle cell centers')
-  }
-
-  console.log(meshData)
-  
-  console.time('assignElevations');
-  const elevationFunction = elevationConfig.useIslands ? assignIslandElevations : assignElevations;
-  const cellElevations = elevationFunction(meshData.cellTriangleCenters, elevationConfig);
-  console.timeEnd('assignElevations');
-
-  console.time('assignBiomes');
-  const cellBiomes: Uint8Array = assignBiomes(
-    meshData.cellTriangleCenters,
-    cellElevations,
-    meshData.cellNeighbors,
-    meshData.cellOffsets,
-    biomeConfig.waterLevel,
-    {
-      frequency: biomeConfig.moistureFrequency,
-      amplitude: biomeConfig.moistureAmplitude,
-      octaves: biomeConfig.moistureOctaves
-    },
-    {
-      frequency: biomeConfig.temperatureFrequency,
-      amplitude: biomeConfig.temperatureAmplitude,
-      octaves: biomeConfig.temperatureOctaves
-    }
-  );
-  currentCellBiomes = cellBiomes;
-  console.timeEnd('assignBiomes');
-
-  // Calculate stats
-  const landCells = cellElevations.filter(e => e >= biomeConfig.waterLevel).length;
-  const waterCells = cellElevations.length - landCells;
-  const landPercentage = Math.round((landCells / cellElevations.length) * 100);
-  
-  // Calculate biome distribution
-  const biomeCounts: { [key: number]: number } = {};
-  for (let i = 0; i < cellBiomes.length; i++) {
-    const biome = cellBiomes[i];
-    biomeCounts[biome] = (biomeCounts[biome] || 0) + 1;
-  }
-  
-  const biomeStatsHtml = Object.entries(biomeCounts)
-    .sort(([,a], [,b]) => b - a)
-    .slice(0, 6) // Show top 6 biomes
-    .map(([biomeId, count]) => {
-      const percentage = Math.round((count / cellBiomes.length) * 100);
-      return `${getBiomeName(parseInt(biomeId))}: ${percentage}%`;
-    })
-    .join('<br>');
-  
-  document.getElementById('stats')!.innerHTML = `
-    Land: ${landPercentage}% (${landCells} cells)<br>
-    Water: ${100 - landPercentage}% (${waterCells} cells)<br>
-    Mesh: ${currentMapSize} (${meshData.cellCount} cells)
-  `;
-  
-  document.getElementById('biomeStats')!.innerHTML = `
-    <strong>Top Biomes:</strong><br>
-    ${biomeStatsHtml}
-  `;
-
-  console.time('drawCells');
-  drawCells(
-    WIDTH,
-    HEIGHT,
-    ctx, 
-    meshData.allVertices, 
-    meshData.cellOffsets, 
-    meshData.cellVertexIndices, 
-    cellBiomes,
-    meshData.cellNeighbors,
-    biomeConfig.smoothColors
-  );
-  console.timeEnd('drawCells');
-}
+const ctx = canvas.getContext('2d')!;
 
 // Initialize the application
 async function initializeApp() {
   console.log('Initializing app...');
-  
+
   // Create UI first
-  createUI();
+  createUI(ctx);
 
   setupGameplayEventListeners();
-  
+
   // Start preloading all meshes in background (non-blocking)
-  preloadAllMeshes(SERVER_BASE_URL);
-  
+  preloadMeshes();
+
   // Load the default mesh (xl)
-  await loadOrGetMesh(currentMapSize);
-  
+  await loadOrGetMesh(currentMapSize, ctx);
+
   console.log('✅ Application initialized');
-}
-
-
-/**
- * ===========================================================================================
- * ======================================== CREATE UI ========================================
- * ===========================================================================================
- */
-
-function createUI() {
-  // Create UI panel
-  const uiPanel = document.createElement("div");
-  uiPanel.style.cssText = `
-    position: fixed;
-    top: 10px;
-    right: 10px;
-    width: 300px;
-    background: rgba(0, 0, 0, 0.8);
-    color: white;
-    padding: 15px;
-    border-radius: 8px;
-    font-family: Arial, sans-serif;
-    font-size: 12px;
-    max-height: 90vh;
-    overflow-y: auto;
-    z-index: 1000;
-  `;
-
-  uiPanel.innerHTML = `
-    <h3 style="margin: 0 0 15px 0; color: #4CAF50;">Biome Terrain Controls</h3>
-    
-    <div style="margin-bottom: 15px;">
-      <label>
-        <input type="checkbox" id="useIslands" ${
-          elevationConfig.useIslands ? "checked" : ""
-        }> 
-        Island Mode
-      </label>
-    </div>
-
-    <div style="margin-bottom: 15px;">
-      <label>
-        <input type="checkbox" id="smoothColors" ${
-          biomeConfig.smoothColors ? "checked" : ""
-        }> 
-        Smooth Colors
-      </label>
-    </div>
-
-    <div style="margin-bottom: 15px;">
-      <label>Map Size:</label>
-      <select id="mapSize" style="width: 100%; margin-top: 5px; background: #333; color: white; border: 1px solid #555; padding: 4px;">
-        <option value="small">Small</option>
-        <option value="medium">Medium</option>
-        <option value="large">Large</option>
-        <option value="xl" selected>XL</option>
-      </select>
-    </div>
-
-    <details style="margin-bottom: 15px;">
-      <summary style="cursor: pointer; margin-bottom: 10px;">Biome Settings</summary>
-      
-      <div style="margin-bottom: 10px;">
-        <label>Water Level: <span id="waterLevelValue">${
-          biomeConfig.waterLevel
-        }</span></label>
-        <input type="range" id="waterLevel" min="0.2" max="0.8" step="0.05" value="${
-          biomeConfig.waterLevel
-        }" 
-               style="width: 100%; margin-top: 5px;">
-      </div>
-
-      <div style="margin-bottom: 10px;">
-        <label>Moisture Frequency: <span id="moistureFrequencyValue">${
-          biomeConfig.moistureFrequency
-        }</span></label>
-        <input type="range" id="moistureFrequency" min="0.005" max="0.05" step="0.005" value="${
-          biomeConfig.moistureFrequency
-        }" 
-               style="width: 100%; margin-top: 5px;">
-      </div>
-
-      <div style="margin-bottom: 10px;">
-        <label>Temperature Frequency: <span id="temperatureFrequencyValue">${
-          biomeConfig.temperatureFrequency
-        }</span></label>
-        <input type="range" id="temperatureFrequency" min="0.005" max="0.05" step="0.005" value="${
-          biomeConfig.temperatureFrequency
-        }" 
-               style="width: 100%; margin-top: 5px;">
-      </div>
-
-      <div style="margin-bottom: 10px;">
-        <label>Moisture Octaves: <span id="moistureOctavesValue">${
-          biomeConfig.moistureOctaves
-        }</span></label>
-        <input type="range" id="moistureOctaves" min="1" max="5" step="1" value="${
-          biomeConfig.moistureOctaves
-        }" 
-               style="width: 100%; margin-top: 5px;">
-      </div>
-
-      <div style="margin-bottom: 15px;">
-        <label>Temperature Octaves: <span id="temperatureOctavesValue">${
-          biomeConfig.temperatureOctaves
-        }</span></label>
-        <input type="range" id="temperatureOctaves" min="1" max="5" step="1" value="${
-          biomeConfig.temperatureOctaves
-        }" 
-               style="width: 100%; margin-top: 5px;">
-      </div>
-      <hr></hr>
-    </details>
-
-    <div style="margin-bottom: 10px;">
-      <label>Elevation Shift: <span id="elevationShiftValue">${
-        elevationConfig.elevationShift
-      }</span></label>
-      <input type="range" id="elevationShift" min="-0.4" max="0.4" step="0.01" value="${
-        elevationConfig.elevationShift
-      }" 
-             style="width: 100%; margin-top: 5px;">
-    </div>
-    
-    <div style="margin-bottom: 10px;">
-      <label>Octaves: <span id="octavesValue">${
-        elevationConfig.octaves
-      }</span></label>
-      <input type="range" id="octaves" min="1" max="6" step="1" value="${
-        elevationConfig.octaves
-      }" 
-             style="width: 100%; margin-top: 5px;">
-    </div>
-    
-    <div style="margin-bottom: 10px;">
-      <label>Redistribution:</label>
-      <select id="redistribution" style="width: 100%; margin-top: 5px; background: #333; color: white; border: 1px solid #555; padding: 4px;">
-        <option value="none">None</option>
-        <option value="linear">Linear</option>
-        <option value="exponential" selected>Exponential</option>
-      </select>
-    </div>
-    
-    <div id="exponentialPowerDiv" style="margin-bottom: 10px;">
-      <label>Exponential Power: <span id="exponentialPowerValue">${
-        elevationConfig.exponentialPower
-      }</span></label>
-      <input type="range" id="exponentialPower" min="0.5" max="3" step="0.1" value="${
-        elevationConfig.exponentialPower
-      }" 
-             style="width: 100%; margin-top: 5px;">
-    </div>
-    
-    <div style="margin-bottom: 10px;">
-      <label>Seed:</label>
-      <div style="display: flex; gap: 5px; margin-top: 5px; align-items: center;">
-        <input type="number" id="seedInput" min="0" max="1" step="0.001" value="${elevationConfig.seed.toFixed(
-          3
-        )}" 
-               style="flex: 1; background: #333; color: white; border: 1px solid #555; padding: 4px; border-radius: 4px;">
-        <button id="randomSeed" style="background: #666; color: white; border: none; padding: 6px 12px; border-radius: 4px; cursor: pointer;">Random</button>
-      </div>
-    </div>
-    
-    <details style="margin-bottom: 10px;">
-      <summary style="cursor: pointer; margin-bottom: 10px;">Advanced (Amplitudes & Frequencies)</summary>
-      
-      <div id="amplitudesContainer">
-        <label>Amplitudes:</label>
-        ${elevationConfig.amplitudes
-          .map(
-            (amp, i) =>
-              `<div style="margin: 5px 0;">
-            <label>Octave ${
-              i + 1
-            }: <span id="amp${i}Value">${amp}</span></label>
-            <input type="range" id="amplitude${i}" min="0" max="1" step="0.025" value="${amp}" style="width: 100%;">
-           </div>`
-          )
-          .join("")}
-      </div>
-      
-      <div id="frequenciesContainer" style="margin-top: 10px;">
-        <label>Frequencies:</label>
-        ${elevationConfig.frequencies
-          .map(
-            (freq, i) =>
-              `<div style="margin: 5px 0;">
-            <label>Octave ${
-              i + 1
-            }: <span id="freq${i}Value">${freq}</span></label>
-            <input type="range" id="frequency${i}" min="0.001" max="0.05" step="0.001" value="${freq}" style="width: 100%;">
-           </div>`
-          )
-          .join("")}
-      </div>
-    </details>
-    
-    <div style="margin-top: 15px; padding-top: 10px; border-top: 1px solid #555; font-size: 11px; color: #aaa;">
-      <div id="stats"></div>
-      <div id="biomeStats" style="margin-top: 10px;"></div>
-    </div>
-
-    <div style="margin-top: 15px; padding-top: 10px; border-top: 1px solid #555; font-size: 11px; color: #aaa;">
-      <div style="display: flex; gap: 10px; margin-bottom: 10px;">
-        <button id="createGame" style="flex: 1; background: #4CAF50; color: white; border: none; padding: 10px; border-radius: 4px; cursor: pointer;">Create Game</button>
-        <button id="joinGame" style="flex: 1; background: #2196F3; color: white; border: none; padding: 10px; border-radius: 4px; cursor: pointer;">Join Game</button>
-      </div>
-    </div>
-  `;
-
-  document.body.appendChild(uiPanel);
-
-  // Map size selectors
-  document.getElementById("mapSize")!.addEventListener("change", async (e) => {
-    const size = (e.target as HTMLSelectElement).value as MapSize;
-    if (size) {
-      currentMapSize = size;
-      await loadOrGetMesh(currentMapSize);
-    }
-  });
-
-  document.getElementById("randomSeed")!.addEventListener("click", () => {
-    elevationConfig.seed = Math.random();
-    (document.getElementById("seedInput") as HTMLInputElement).value =
-      elevationConfig.seed.toFixed(3);
-    generateTerrain();
-  });
-
-  // Seed input
-  document.getElementById("seedInput")!.addEventListener("input", (e) => {
-    const value = parseFloat((e.target as HTMLInputElement).value);
-    if (!isNaN(value)) {
-      elevationConfig.seed = Math.max(0, Math.min(1, value));
-      generateTerrain();
-    }
-  });
-
-  // Range inputs - all auto-regenerate
-  ["elevationShift", "octaves", "exponentialPower"].forEach((param) => {
-    const element = document.getElementById(param)!;
-    element.addEventListener("input", (e) => {
-      const value = parseFloat((e.target as HTMLInputElement).value);
-      (elevationConfig as any)[param] =
-        param === "octaves" ? Math.floor(value) : value;
-      document.getElementById(param + "Value")!.textContent = value.toString();
-      generateTerrain();
-    });
-  });
-
-  // Biome config inputs
-  [
-    "waterLevel",
-    "moistureFrequency",
-    "temperatureFrequency",
-    "moistureOctaves",
-    "temperatureOctaves",
-  ].forEach((param) => {
-    const element = document.getElementById(param)!;
-    element.addEventListener("input", (e) => {
-      const value = parseFloat((e.target as HTMLInputElement).value);
-      (biomeConfig as any)[param] = [
-        "moistureOctaves",
-        "temperatureOctaves",
-      ].includes(param)
-        ? Math.floor(value)
-        : value;
-      document.getElementById(param + "Value")!.textContent = value.toString();
-      generateTerrain();
-    });
-  });
-
-  // Redistribution
-  document.getElementById("redistribution")!.addEventListener("change", (e) => {
-    elevationConfig.redistribution = (e.target as HTMLSelectElement)
-      .value as any;
-    document.getElementById("exponentialPowerDiv")!.style.display =
-      elevationConfig.redistribution === "exponential" ? "block" : "none";
-    generateTerrain();
-  });
-
-  // Islands checkbox
-  document.getElementById("useIslands")!.addEventListener("change", (e) => {
-    elevationConfig.useIslands = (e.target as HTMLInputElement).checked;
-    generateTerrain();
-  });
-
-  // Smooth colors checkbox
-  document.getElementById("smoothColors")!.addEventListener("change", (e) => {
-    biomeConfig.smoothColors = (e.target as HTMLInputElement).checked;
-    generateTerrain(); // Re-render with new color smoothing
-  });
-
-  // Amplitude and frequency controls - auto-regenerate
-  for (let i = 0; i < 4; i++) {
-    const ampElement = document.getElementById(`amplitude${i}`);
-    const freqElement = document.getElementById(`frequency${i}`);
-
-    if (ampElement) {
-      ampElement.addEventListener("input", (e) => {
-        const value = parseFloat((e.target as HTMLInputElement).value);
-        elevationConfig.amplitudes[i] = value;
-        document.getElementById(`amp${i}Value`)!.textContent = value.toString();
-        generateTerrain();
-      });
-    }
-
-    if (freqElement) {
-      freqElement.addEventListener("input", (e) => {
-        const value = parseFloat((e.target as HTMLInputElement).value);
-        elevationConfig.frequencies[i] = value;
-        document.getElementById(`freq${i}Value`)!.textContent =
-          value.toString();
-        generateTerrain();
-      });
-    }
-  }
 }
 
 // Utility function to manage game button states
@@ -1534,16 +1103,16 @@ function processGameData(gameData: any): void {
     console.log(`Current player: ${gameData.state.currentPlayer}, My turn: ${isMyTurn}`);
     
     // Update the current biomes with the game terrain
-    currentCellBiomes = terrainData;
-    currentCellCount = terrainData.length;
-    
+    setCurrentCellBiomes(terrainData);
+    setCurrentCellCount(terrainData.length);
+
     // Update the map size to match the game
-    currentMapSize = gameData.meta.mapSize as MapSize;
+    setCurrentMapSize(gameData.meta.mapSize as MapSize);
     
     // Load the appropriate mesh if we don't have it
     if (!meshData || meshData.cellCount !== terrainData.length) {
       console.log(`Loading ${gameData.meta.mapSize} mesh for game...`);
-      loadOrGetMesh(gameData.meta.mapSize as MapSize).then(() => {
+      loadOrGetMesh(gameData.meta.mapSize as MapSize, ctx).then(() => {
         renderGameState();
       });
     } else {
@@ -1675,72 +1244,7 @@ let currentGameId: string | null = null;
 let currentPlayerName: string | null = null;
 let isGameCreator: boolean = false;
 
-let currentCellBiomes: Uint8Array = new Uint8Array(0);
-let currentCellCount: number = 0;
-
-// TODO we need to standardize/centralize this configuration between client and server
-const WIDTH  = 960;
-const HEIGHT = 600;
-
-// Server configuration
-const SERVER_BASE_URL: string = import.meta.env.VITE_SERVER_URL || 'http://localhost:3000';
-
-// Biome color scheme
-export const BIOME_COLORS: { [key: number]: string } = {
-  0: "#88aa55",   // Plains - Grassland
-  1: "#679459",   // Woods - Temperate Deciduous Forest
-  2: "#337755",   // Rainforest - Tropical Rain Forest
-  3: "#2f6666",   // Wetlands - Marsh
-  4: "#889977",   // Hills - Shrubland
-  5: "#888888",   // Mountains - Bare
-  6: "#44447a",   // Shallow Ocean - Ocean
-  7: "#33335a",   // Deep Ocean - (darker)
-  8: "#bbbbaa",   // Tundra Plains - Tundra
-  9: "#99aa77",   // Tundra Woods - Taiga
-  10: "#bbbbaa",  // Tundra Hills - Tundra
-  11: "#ffffff",  // Tundra Mountains - Ice
-  12: "#d2b98b",  // Desert Plains - Subtropical Desert
-  13: "#c9d29b",  // Desert Hills - Temperate Desert
-  14: "#555555"   // Desert Mountains - Scorched
-};
-
-// set up canvas
-const canvas = document.createElement('canvas');
-const container = document.getElementById('canvas-container')!;
-container.appendChild(canvas);
-canvas.width  = WIDTH;
-canvas.height = HEIGHT;
-
-const ctx = canvas.getContext('2d')!;
-
-// Current mesh data
-let meshData: MeshData | null = null;
-let currentMapSize: MapSize = 'xl';
-
-// Current elevation configuration
-let elevationConfig: ElevationConfig = {
-  amplitudes: [0.6, 0.3, 0.15, 0.075],
-  frequencies: [0.003, 0.006, 0.012, 0.024],
-  octaves: 4,
-  seed: Math.random(),
-  waterLevel: 0.5,
-  redistribution: 'exponential' as const,
-  exponentialPower: 1.5,
-  elevationShift: -0.1,
-  useIslands: false
-};
-
-// Biome configuration
-let biomeConfig: BiomeConfig = {
-  waterLevel: 0.5,
-  moistureFrequency: 0.02,
-  moistureAmplitude: 1.0,
-  moistureOctaves: 3,
-  temperatureFrequency: 0.015,
-  temperatureAmplitude: 1.0,
-  temperatureOctaves: 2,
-  smoothColors: true
-};
+// Terrain configuration and canvas setup moved to separate modules
 
 // Start the application
 initializeApp().catch(error => {
